Fix possible race condition in scheduler
