+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io/kotest/engine/TestSuiteScheduler.kt
@@ -1,27 +1,22 @@
-@file:Suppress("DEPRECATION")
-
 package io.kotest.engine
 
 import io.kotest.core.Logger
 import io.kotest.core.Platform
-import io.kotest.core.annotation.DoNotParallelize
 import io.kotest.core.annotation.Isolate
 import io.kotest.core.annotation.Parallel
 import io.kotest.core.platform
 import io.kotest.core.project.TestSuite
 import io.kotest.core.spec.SpecRef
+import io.kotest.engine.concurrency.isIsolate
+import io.kotest.engine.concurrency.isParallel
 import io.kotest.engine.interceptors.EngineContext
 import io.kotest.engine.listener.CollectingTestEngineListener
 import io.kotest.engine.spec.SpecExecutor
-import io.kotest.mpp.IncludingAnnotations
-import io.kotest.mpp.IncludingSuperclasses
 import io.kotest.mpp.bestName
-import io.kotest.mpp.hasAnnotation
 import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Semaphore
 import kotlinx.coroutines.sync.withPermit
-import kotlin.reflect.KClass
 
 /**
  * A [TestSuiteScheduler] schedules specs for execution using a [SpecExecutor].
@@ -39,16 +34,19 @@ internal class TestSuiteScheduler(
    suspend fun schedule(suite: TestSuite): EngineResult {
       logger.log { Pair(null, "Launching ${suite.specs.size} specs") }
 
+      // first we run the specs that have been marked as always isolated
       val isolated = suite.specs.filter { it.kclass.isIsolate() }
       logger.log { Pair(null, "Isolated spec count: ${isolated.size}") }
       schedule(isolated, 1)
       logger.log { Pair(null, "Isolated specs have completed") }
 
+      // first we run the specs that have been marked as always parallel regardless of concurrency mode
       val parallel = suite.specs.filter { it.kclass.isParallel() }
       logger.log { Pair(null, "Parallelized spec count: ${parallel.size}") }
       schedule(parallel, Int.MAX_VALUE)
       logger.log { Pair(null, "Parallelized specs have completed") }
 
+      // the rest of the specs use the concurrency mode
       val default = suite.specs.filter { !it.kclass.isIsolate() && !it.kclass.isParallel() }
       logger.log { Pair(null, "Remaining spec count: ${default.size}") }
       schedule(default, concurrency())
@@ -62,7 +60,9 @@ internal class TestSuiteScheduler(
       concurrency: Int,
    ) {
 
+      // we use this to check for failures for fast failure mode
       val collector = CollectingTestEngineListener()
+      val mergedContext = context.mergeListener(collector)
 
       val semaphore = Semaphore(concurrency)
       logger.log { Pair(null, "Scheduling using concurrency: $concurrency") }
@@ -73,22 +73,30 @@ internal class TestSuiteScheduler(
             launch {
                semaphore.withPermit {
                   logger.log { Pair(ref.kclass.bestName(), "Acquired permit") }
-                  if (context.projectConfigResolver.projectWideFailFast() && collector.errors) {
-                     logger.log { Pair(ref.kclass.bestName(), "Project wide fail fast is active, skipping spec") }
-                     context.listener.specIgnored(ref.kclass, null)
-                  } else {
-                     try {
-                        val executor = SpecExecutor(context.mergeListener(collector))
-                        logger.log { Pair(ref.kclass.bestName(), "Executing ref") }
-                        executor.execute(ref)
-                     } catch (t: Throwable) {
-                        logger.log { Pair(ref.kclass.bestName(), "Unhandled error during spec execution $t") }
-                        throw t
-                     }
-                  }
+                  executeIfNotFailedFast(mergedContext, ref, collector)
                }
                logger.log { Pair(ref.kclass.bestName(), "Released permit") }
             }
+         }
+      }
+   }
+
+   private suspend fun executeIfNotFailedFast(
+      context: EngineContext,
+      ref: SpecRef,
+      collector: CollectingTestEngineListener,
+   ) {
+      if (context.projectConfigResolver.projectWideFailFast() && collector.errors) {
+         logger.log { Pair(ref.kclass.bestName(), "Project wide fail fast is active, skipping spec") }
+         context.listener.specIgnored(ref.kclass, null)
+      } else {
+         try {
+            val executor = SpecExecutor(context)
+            logger.log { Pair(ref.kclass.bestName(), "Executing ref") }
+            executor.execute(ref)
+         } catch (t: Throwable) {
+            logger.log { Pair(ref.kclass.bestName(), "Unhandled error during spec execution $t") }
+            throw t
          }
       }
    }
@@ -107,20 +115,3 @@ internal class TestSuiteScheduler(
       }
    }
 }
-
-/**
- * Returns true if this class is annotated with either of the annotations used to indicate
- * this spec should not run concurrently regardless of config.
- *
- * Those annotations are [DoNotParallelize] and [Isolate].
- */
-@Suppress("DEPRECATION")
-internal fun KClass<*>.isIsolate(): Boolean =
-   hasAnnotation<DoNotParallelize>(IncludingAnnotations, IncludingSuperclasses)
-      || hasAnnotation<Isolate>(IncludingAnnotations, IncludingSuperclasses)
-
-/**
- * Returns true if this class is annotated with the annotation used to indicate
- * this spec should always run concurrently regardless of config.
- */
-internal fun KClass<*>.isParallel() = hasAnnotation<Parallel>(IncludingAnnotations, IncludingSuperclasses)
